Don't include the Gemfile.lock in the gem.
If a dependency installs one of its dependencies when the backstage gem
is installed, it can result in a locked gem not being available. Without
the lock file, bundler will lock when first invoked, and should see all
the gems it needs since they were installed by rubygems' dependency
mechanism.